ui system work

+++ src/Headers/Rect.h
@@ -4,12 +4,14 @@
 struct Rect {
   int x;
   int y;
-  int width;
-  int height;
+  unsigned int width;
+  unsigned int height;
 
   bool overlaps(const Rect &other) const {
-    return x < other.x + other.width && x + width > other.x &&
-           y < other.y + other.height && y + height > other.y;
+    return x < other.x + static_cast<int>(other.width) &&
+           x + static_cast<int>(width) > other.x &&
+           y < other.y + static_cast<int>(other.height) &&
+           y + static_cast<int>(height) > other.y;
   }
 };
